requirements-web.txt
flask==3.1.2
pandas==2.1.0
numpy==1.26.0
matplotlib==3.10.7
scipy==1.13.0
seaborn==0.13.2
google-generativeai==0.8.5
immutabledict==4.2.2
pillow==12.0.0
dataclasses-json==0.6.7
termcolor==3.1.0
aenum==3.1.16
absl-py==2.3.1
jinja2==3.1.6
pyyaml==6.0.3
nbformat==5.10.4
tavily-python==0.7.12
python-dotenv==1.0.0
gunicorn==21.2.0
